Convert app to Vercel serverless
