+++ package.json
@@ -4,20 +4,15 @@
   "description": "Interactive pipeline visualization for Affinity CRM leads",
   "main": "index.js",
   "scripts": {
-    "start": "node server.js",
-    "dev": "nodemon server.js"
+    "start": "vercel dev",
+    "dev": "vercel dev"
   },
   "dependencies": {
-    "express": "^4.18.2",
     "axios": "^1.6.0",
-    "cors": "^2.8.5",
     "dotenv": "^16.3.1",
     "d3": "^7.8.5"
-  },
-  "devDependencies": {
-    "nodemon": "^3.0.1"
   },
   "keywords": ["affinity", "crm", "pipeline", "visualization"],
   "author": "",
   "license": "MIT"
-} 
+}
