+++ public/index.html
@@ -12,13 +12,6 @@
         <!-- Header -->
         <header class="header">
             <h1><i class="fas fa-chart-line"></i> Affinity Pipeline Visualization</h1>
-            <div class="header-controls">
-                <div class="api-key-section">
-                    <label for="apiKey">API Key:</label>
-                    <input type="password" id="apiKey" placeholder="Enter your Affinity API key">
-                    <button id="saveApiKey" class="btn btn-primary">Save</button>
-                </div>
-            </div>
         </header>
 
         <!-- Configuration Panel -->
